requirements.txt
# txtFlask
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
PyMySQL==1.1.0
requests==2.31.0
numpy==1.24.3
plotly==5.15.0
Werkzeug==2.3.7